requirements.txt
python-gitlab>=4.3.0,<5